[Fix] #128 공통 파일 수정
