--- momogum/Card/OtherCard/OtherCardModel.swift
@@ -1,14 +1,0 @@
-//
-//  OtherCardModel.swift
-//  momogum
-//
-//  Created by 조승연 on 2/5/25.
-//
-
-import Foundation
-
-struct OtherCardModel {
-    var likeCount: Int
-    var reviewText: String
-    var showBookmark: Bool
-}
